Fix the linux modules build.  Comparision between signed and unsigned integers. Review URL: http://codereview.chromium.org/28232
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@10536 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/extensions/extension.cc
@@ -69,7 +69,7 @@ const char* Extension::kInvalidZipHashError =
 const char* Extension::kInvalidPluginsDirError =
     "Invalid value for 'plugins_dir'.";
 
-const int Extension::kIdSize = 20;  // SHA1 (160 bits) == 20 bytes
+const size_t Extension::kIdSize = 20;  // SHA1 (160 bits) == 20 bytes
 
 const std::string Extension::VersionString() const {
   return version_->GetString();
